Issue#183: if deliveryMaterial is deleted, the same material on movie is deleted
