+++ apps/backend-functions/src/notify.ts
@@ -20,6 +20,15 @@ interface Notification extends PartialNotification {
   date: any;
 }
 
+interface SnapObject {
+  movieTitle: string,
+  docID: DocID,
+  stakeholderId: string,
+  orgName: string,
+  count: number,
+  userIds: string[]
+}
+
 export async function triggerNotifications(notifications: Notification[]): Promise<any> {
   const notificationBatch = db.batch();
 
@@ -38,4 +47,19 @@ export function prepareNotification(notif: PartialNotification): Notification {
     date: serverTimestamp(),
     ...notif
   } as Notification;
+}
+
+export function customMessage(userId: string, snap: SnapObject) {
+  if (snap.docID.type === 'delivery') {
+    return snap.userIds.includes(userId) && snap.count > 1
+      ? `You have been invited to work on ${snap.movieTitle}'s ${snap.docID.type}. Do you wish to work on it ?`
+      : `${snap.orgName} has been added to ${snap.movieTitle}'s ${snap.docID.type}`;
+    }
+  if (snap.docID.type === 'movie') {
+    return snap.userIds.includes(userId) && snap.count > 1
+      ? `You have been invited to work on ${snap.movieTitle}. Do you wish to work on it ?`
+      : `${snap.orgName} has been added to ${snap.movieTitle}`;
+  } else {
+    throw new Error('Message is not valid');
+  }
 }
